# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Tekton Hub
    description: HTTP services for managing Tekton Hub
    version: "1.0"
servers:
    - url: https://api.hub.tekton.dev
paths:
...
    /resource/{catalog}/{kind}/{name}:
        get:
            tags:
                - resource
            summary: ByCatalogKindName resource
            description: Find resources using name of catalog, resource name and kind of resource
            operationId: resource#ByCatalogKindName
            parameters:
                - name: pipelinesversion
                  in: query
                  description: To find resource compatible with a Tekton pipelines version, use this param
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: To find resource compatible with a Tekton pipelines version, use this param
                    example: 0.21.0
                    pattern: ^\d+(?:\.\d+){0,2}$
                  example: 0.21.0
                - name: catalog
                  in: path
                  description: name of catalog
                  required: true
                  schema:
                    type: string
                    description: name of catalog
                    example: tektoncd
                  example: tektoncd
                - name: kind
                  in: path
                  description: kind of resource
                  required: true
                  schema:
                    type: string
                    description: kind of resource
                    example: pipeline
                    enum:
                        - task
                        - pipeline
                  example: task
                - name: name
                  in: path
                  description: Name of resource
                  required: true
                  schema:
                    type: string
                    description: Name of resource
                    example: buildah
                  example: buildah
            responses:
                "302":
                    description: Found response.
                    headers:
                        location:
                            description: Redirect URL
                            required: true
                            schema:
                                type: string
                                description: Redirect URL
                                example: http://waterslemke.org/mabelle
                                format: uri
                            example: http://ohara.name/vicente_rogahn
    /resource/{catalog}/{kind}/{name}/{version}:
        get:
            tags:
                - resource
            summary: ByCatalogKindNameVersion resource
            description: Find resource using name of catalog & name, kind and version of resource
            operationId: resource#ByCatalogKindNameVersion
            parameters:
                - name: catalog
                  in: path
                  description: name of catalog
                  required: true
                  schema:
                    type: string
                    description: name of catalog
                    example: tektoncd
                  example: tektoncd
                - name: kind
                  in: path
                  description: kind of resource
                  required: true
                  schema:
                    type: string
                    description: kind of resource
                    example: task
                    enum:
                        - task
                        - pipeline
                  example: task
                - name: name
                  in: path
                  description: name of resource
                  required: true
                  schema:
                    type: string
                    description: name of resource
                    example: buildah
                  example: buildah
                - name: version
                  in: path
                  description: version of resource
                  required: true
                  schema:
                    type: string
                    description: version of resource
                    example: "0.1"
                  example: "0.1"
            responses:
                "302":
                    description: Found response.
                    headers:
                        location:
                            description: Redirect URL
                            required: true
                            schema:
                                type: string
                                description: Redirect URL
                                example: http://mantehegmann.com/ward_towne
                                format: uri
                            example: http://koch.net/waino
...
    /resource/{id}/versions:
        get:
            tags:
                - resource
            summary: VersionsByID resource
            description: Find all versions of a resource by its id
            operationId: resource#VersionsByID
            parameters:
                - name: id
                  in: path
                  description: ID of a resource
                  required: true
                  schema:
                    type: integer
                    description: ID of a resource
                    example: 1
                  example: 1
            responses:
                "302":
                    description: Found response.
                    headers:
                        location:
                            description: Redirect URL
                            required: true
                            schema:
                                type: string
                                description: Redirect URL
                                example: http://fahey.org/elta
                                format: uri
                            example: http://bayer.name/lafayette
    /resource/version/{versionID}:
        get:
            tags:
                - resource
            summary: ByVersionId resource
            description: Find a resource using its version's id
            operationId: resource#ByVersionId
            parameters:
                - name: versionID
                  in: path
                  description: Version ID of a resource's version
                  required: true
                  schema:
                    type: integer
                    description: Version ID of a resource's version
                    example: 1
                  example: 1
            responses:
                "302":
                    description: Found response.
                    headers:
                        location:
                            description: Redirect URL
                            required: true
                            schema:
                                type: string
                                description: Redirect URL
                                example: http://christiansen.name/haylee
                                format: uri
                            example: http://connellyschaefer.biz/elody_schumm
...
components:
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Secures endpoint by requiring a valid JWT retrieved via the /auth/login endpoint.
            scheme: bearer
...